BaseRunner ABC + BlackRhino
Added BaseRunner class, and moved BlackRhino to conform with the 4 ABCs,
also infocontagion works with the 4th ABC, still need to work on
streamlining it all more over the coming months, together with another
ABM (DSGE one), to make them all more homogenous. But for now BlackRhino
should be operational in this version, tests work.